ct/l0/gc: Hoist epoch_source into a separate module
We will need it in the barrier as well as the garbage collector.

Only minor changes:
- add epoch_source::make_default
- namespace fiddling
